Refactor to Task updates to services
-> Refactor update_tasks_with_file_annotations() and get_file_with_annotations() to shared services.
-> Reuse in task list
-> Add to annotator dashboard (was missing). Fixes missing preview images and counts
-> Remove get_file_with_annotations() it's not used any more
