+++ frontend/src/components/task/task/task_list.vue
@@ -369,12 +369,12 @@ import {
   assignUserToTask,
   batchAssignUserToTask,
   batchRemoveUserFromTask,
-  getTaskListFromJob
+  getTaskListFromJob,
+  update_tasks_with_file_annotations
 } from "../../../services/tasksServices.js"
+
 import {get_task_template_members} from "../../../services/taskTemplateService.js"
 import global_dataset_selector from "../../attached/global_dataset_selector.vue"
-
-import pLimit from "p-limit";
 
 import Vue from "vue";
 
@@ -856,7 +856,7 @@ export default Vue.extend({
             this.column_list_all = this.column_list_all.filter(col => !["AssignedReviewer"].includes(col))
           }
 
-          this.update_tasks_with_file_annotations(this.task_list);
+          update_tasks_with_file_annotations(this.task_list);
         }
         return task_data;
       } catch (error) {
@@ -867,39 +867,6 @@ export default Vue.extend({
       }
     },
 
-    update_tasks_with_file_annotations: async function (task_list) {
-      const limit = pLimit(10); // Max concurrent request.
-      try {
-        const promises = task_list.map((task) => {
-          return limit(() => this.get_file_with_annotations(task));
-        });
-        const result = await Promise.all(promises);
-        return result;
-      } catch (error) {
-        this.file_update_error = this.$route_api_errors(error);
-        console.error(error);
-      }
-    },
-
-    async get_file_with_annotations(task) {
-      if (task.file.type != "image") {
-        return;
-      }
-
-      let url = "/api/v1/task/" + task.id + "/annotation/list";
-      this.get_annotations_error = {};
-      this.get_annotations_loading = true;
-
-      try {
-        const response = await axios.post(url, {});
-        task.file = response.data.file_serialized;
-      } catch (error) {
-        console.debug(error);
-        this.get_annotations_error = this.$route_api_errors(error);
-      } finally {
-        this.get_annotations_loading = false;
-      }
-    },
 
     async refresh_task_list() {
       this.loading = true;
